fix(aero-machine): align vbe_lfb_base test with u64 API
Machine::vbe_lfb_base() now returns u64 (matching other physical address helpers),
while VBE itself stores PhysBasePtr as a u32. Cast the firmware default LFB
base constant in the unit test so the assertion compiles.

+++ crates/aero-machine/src/lib.rs
@@ -12577,7 +12577,7 @@ mod tests {
         .unwrap();
         assert_eq!(
             headless.vbe_lfb_base(),
-            firmware::video::vbe::VbeDevice::LFB_BASE_DEFAULT
+            u64::from(firmware::video::vbe::VbeDevice::LFB_BASE_DEFAULT)
         );
 
         // When VGA is enabled, the BIOS should report the legacy MMIO-mapped SVGA base.
